When built as flatpak change freedesktop prefix
RStudio installs freedesktop files into /usr/... which is the right
place when being installed on ordinary systems. However, flatpaks
need these files to be placed in /app/... because /usr and such are
read-only.

As a solution I created a variable that is set to /usr or /app depending
on a Flatpak flag being present.